Move PopupMessage type into separate file.
